requirements.txt
# general  modules
django >3.2,<3.3
djangorestframework
djangorestframework-simplejwt
yfinance
django-cors-headers
python-decouple

# deployment module
gunicorn
